tools/tidal-dl-helper-scripts/requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
tidalapi>=0.8.0
